+++ docs/CHANGELOG.rst
@@ -60,6 +60,8 @@ Changed
 Fixed
 -----
 * Fix ``consensus_subreads`` input option in Subread process
+* Fix ``cuffnorm`` process to correctly use sample names as labels in
+  output files and expand ``cuffnorm`` tests
 
 
 ==================